Fix #19 OOM + empty storage browser (v0.6.0)
Bounded subvolume inference and storage-browser fixes so the hosted
demo runs within the web container's memory and the browser lists
real ADLS contents.

#19 (OOM in live fault detection):
- InterpretationRequest: add inline_center/inline_window; None = full cube.
- _run_fault_detection: slice a +/-window inline slab (>= one patch thick),
  run inference on it, and record inline/crossline/twtt coords in the run
  manifest.
- VolumeInference.run: float32 accumulators (was float64) — halves peak RAM.
- get_overlay: accept an ABSOLUTE inline and map to the local volume index
  via the manifest coords; return real crossline/twtt from the manifest.
- UNet3D.save_checkpoint: recursively sanitize metrics to JSON-safe scalars
  so Windows-saved checkpoints (WindowsPath in train_config) load on Linux.

Empty storage browser:
- browse route: hierarchical walk_blobs(delimiter="/") instead of a flat
  list over the whole (multi-TB) container, which timed out -> empty UI.
- _viewer_api.browse(): fail-loud client; gradio browse surfaces errors.

UI:
- vapi.start_fault_detection sends inline_center/window; gradio passes the
  current inline as center and the absolute inline to fetch_overlay.

Tests: +9 (browse fail-loud, subvolume start params, absolute-inline overlay
mapping + out-of-window 404, checkpoint OS-portability). 323 pass, ruff clean.

+++ src/deepseismic/models/inference.py
@@ -239,8 +239,11 @@ class VolumeInference:
                     "Reduce patch_size or use a larger volume."
                 )
 
-        prob_acc   = np.zeros(vol_shape, dtype=np.float64)
-        weight_acc = np.zeros(vol_shape, dtype=np.float64)
+        # float32 accumulators — float64 would double the footprint and is the
+        # primary OOM cause on a full cube (issue #19): for ST10010
+        # [401x720x850] two float64 buffers alone are ~3.9 GB.
+        prob_acc   = np.zeros(vol_shape, dtype=np.float32)
+        weight_acc = np.zeros(vol_shape, dtype=np.float32)
 
         all_windows = list(_sliding_windows(vol_shape, self.patch_size, self.overlap))
         n_patches   = len(all_windows)
